feat: add run history to profile
