+++ src/navigation/tab-navigator.tsx
@@ -50,6 +50,26 @@ const tabsIcons: TabIconsType = {
   // Settings: (props: SvgProps) => <SettingsIcon {...props} />,
 };
 
+import { StackActions } from '@react-navigation/native';
+
+// from https://stackoverflow.com/a/67895977
+const resetStacksOnTabClicks = ({ navigation }: any) => ({
+  tabPress: (e: any) => {
+    const state = navigation.getState();
+    if (state) {
+      const nonTargetTabs = state.routes.filter((r: any) => r.key !== e.target);
+      nonTargetTabs.forEach((tab: any) => {
+        if (tab?.state?.key) {
+          navigation.dispatch({
+            ...StackActions.popToTop(),
+            target: tab?.state?.key,
+          });
+        }
+      });
+    }
+  },
+});
+
 export type TabList<T extends keyof TabParamList> = {
   navigation: NativeStackNavigationProp<TabParamList, T>;
   route: RouteProp<TabParamList, T>;
@@ -101,8 +121,9 @@ export const TabNavigator = () => {
     const userRef = doc(db, 'users', currentUserId).withConverter(
       userConverter
     );
-    const unsubscribe = onSnapshot(userRef, (snapshot) => {
+    const unsubscribe = onSnapshot(userRef, async (snapshot) => {
       const data = snapshot.data();
+
       if (data) {
         setUser({
           ...data,
@@ -143,6 +164,7 @@ export const TabNavigator = () => {
                 title: label,
                 tabBarTestID: `${name}-tab`,
               }}
+              listeners={resetStacksOnTabClicks}
             />
           );
         })}
