+++ src/screens/new_run/index.tsx
@@ -10,8 +10,9 @@ import { useAuth } from '@/core';
 import { FocusAwareStatusBar, Text, TouchableOpacity, View } from '@/ui';
 
 import { Run } from '../run';
-import { RunReport } from '../run_report';
+import { RunReportModal } from '../run_report/run-report-modal';
 
+/* eslint-disable max-lines-per-function */
 export const NewRun: React.FC = () => {
   // state to control modal visibility
   const [friends, setFriends] = useState<User[]>([]);
@@ -107,7 +108,7 @@ export const NewRun: React.FC = () => {
           // onRequestClose={handlePress} // for Android back button
         >
           <View>
-            <RunReport
+            <RunReportModal
               runId={runReportId!}
               onFinish={() => {
                 setRunReportId(null);
