+++ src/screens/profile/index.tsx
@@ -4,6 +4,8 @@ import React from 'react';
 import { useAuth } from '@/core';
 import { FocusAwareStatusBar, Image, ScrollView, View } from '@/ui';
 
+import { RunHistory } from './run-history';
+
 const EmptyState = () => {
   return (
     <View className="flex-1 pt-4">
@@ -62,9 +64,11 @@ export const Profile = () => {
 
           <TabView value={index} onChange={setIndex} animationType="spring">
             <TabView.Item>
-              <ScrollView className="px-4">
-                <Text h1>Current User</Text>
-                <Text>{JSON.stringify(user)}</Text>
+              <ScrollView className="w-screen px-4 pt-4">
+                <Text h4>Your Run History</Text>
+                <View className="min-h-20">
+                  <RunHistory user={user} />
+                </View>
               </ScrollView>
             </TabView.Item>
             <TabView.Item>
