+++ src/screens/run_report/run-report-modal.tsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import { SafeAreaView } from 'react-native';
+
+import { ModalHeader } from '@/ui/core/modal/modal-header';
+
+import { RunReport } from './run-report';
+
+type RunReportModalProps = {
+  runId: string;
+  onFinish: () => void;
+};
+
+export const RunReportModal = ({ runId, onFinish }: RunReportModalProps) => {
+  return (
+    <SafeAreaView>
+      <ModalHeader
+        title=""
+        dismiss={() => {
+          onFinish();
+        }}
+      />
+      <RunReport runId={runId} />
+    </SafeAreaView>
+  );
+};
