+++ src/screens/run_report/run-report-screen.tsx
@@ -1,0 +1,17 @@
+import type { NativeStackScreenProps } from '@react-navigation/native-stack';
+import React from 'react';
+
+import type { ProfileStackParamList } from '@/navigation/profile-navigator';
+
+import { RunReport } from './run-report';
+
+// from https://reactnavigation.org/docs/typescript/
+type RunReportScreenProps = NativeStackScreenProps<
+  ProfileStackParamList,
+  'RunReport'
+>;
+
+export const RunReportScreen = ({ route }: RunReportScreenProps) => {
+  const { runId } = route.params;
+  return <RunReport runId={runId} />;
+};
