+++ src/ui/utils.tsx
@@ -46,3 +46,27 @@ export const extractError = (data: unknown): string => {
   }
   return 'Something went wrong ';
 };
+
+// Turns a unix epoch timestamp into relative time string
+export function timeSince(timeStamp: number) {
+  var now = new Date(),
+    secondsPast = Math.round((now.getTime() - timeStamp) / 1000);
+  if (secondsPast < 60) {
+    return secondsPast + 's';
+  }
+  if (secondsPast < 3600) {
+    return `${Math.floor(secondsPast / 60)} min`;
+  }
+  if (secondsPast <= 86400) {
+    return `${Math.floor(secondsPast / 3600)} h`;
+  }
+  if (secondsPast <= 2628000) {
+    return `${Math.floor(secondsPast / 86400)} day`;
+  }
+  if (secondsPast <= 31536000) {
+    return `${Math.floor(secondsPast / 2628000)} mo`;
+  }
+  if (secondsPast > 31536000) {
+    return `${Math.floor(secondsPast / 31536000)} y`;
+  }
+}
